Add lib, make builds from the project root
Changing the printf string in foo.c then running make from the root
does find the change, build, and install the lib. Then the root source
file is linked with the new lib and the change in the string is there.

This is what I was trying to test, as it has not been working in
another project.